// client-id/src/app/core/interfaces/test-complete-status-rp.ts
export interface TestCompleteSttRp{
    athrzSn: number;
    athrzRceptNo: string;
    athrzReqstNo: number;
    fomConfmConfmNo: string;
    mrnrSeNm: string;
    mrnrSeCnt: number;
    reqstQy: number;
    rqstDt: Date;
    athrzSeCode: string;
    athrzSeCodeName: string;
    vesslNoStr: string;
    entrpsNm: string;
    athrzProcessSttusName: string;
    athrzProcessSttusCode: string;
    statusClassName2: string;
    cnfrmnIssuPrintStr: string;
    rceptConfmDe: Date;
    rceptChargerId: string;
    rceptChargerNm: string;
    rceptConfmerId: string;
    rceptConfmerNm: string;
    athrzChargerNm: string;
    athrzEndDe: string;
    athrzHopeDe: string;
    applcntNm: string;
    insttNm: string;
    lastConfmDe: string;
    rowNumber: number;
    totalCount: number;
}